refactor: Switch to second-based delta

+++ app/src/main/java/io/github/bfur64/tetrue/game/GameManager.java
@@ -68,7 +68,7 @@ public class GameManager {
         new SoundSystem(eventBus);
     }
 
-    public void update(long delta) {
+    public void update(double delta) {
         onGroundSystem.update();
         gravityTimerSystem.update(delta);
         lockTimerSystem.update(delta);
