+++ app/src/main/java/io/github/bfur64/tetrue/game/event/GravityChangeEvent.java
@@ -1,3 +1,3 @@
 package io.github.bfur64.tetrue.game.event;
 
-public record GravityChangeEvent(int gravity) {}
+public record GravityChangeEvent(double gravity) {}
